BlImlp

+++ dotNet5781_ 2375_6415/BL/BO/PairStations.cs
@@ -10,7 +10,7 @@ namespace BO
     {
         public int FirstStationNumber { get; set; }
         public int LastStationNumber { get; set; }
-        public float Distance { get; set; }
+        public double Distance { get; set; }
         public TimeSpan Time { get; set; }
     }
 }
